rename 'fromOrdToDoc' to 'all'; move fromIndexedDISI to codecs/lucene90
