+++ lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/lucene92/OffHeapFloatVectorValues.java
@@ -172,7 +172,7 @@ abstract class OffHeapFloatVectorValues extends FloatVectorValues {
 
     @Override
     protected DocIndexIterator createIterator() {
-      return DocIndexIterator.fromIndexedDISI(disi);
+      return IndexedDISI.asDocIndexIterator(disi);
     }
 
     @Override
