+++ lucene/core/src/java/org/apache/lucene/codecs/lucene90/IndexedDISI.java
@@ -18,6 +18,8 @@ package org.apache.lucene.codecs.lucene90;
 
 import java.io.DataInput;
 import java.io.IOException;
+
+import org.apache.lucene.index.KnnVectorValues;
 import org.apache.lucene.search.DocIdSetIterator;
 import org.apache.lucene.store.IndexInput;
 import org.apache.lucene.store.IndexOutput;
@@ -438,6 +440,40 @@ public final class IndexedDISI extends DocIdSetIterator {
 
   // ALL variables
   int gap;
+
+  /**
+   * Returns an iterator that delegates to the IndexedDISI. Advancing this iterator will advance
+   * the underlying IndexedDISI, and vice-versa.
+   */
+  public static KnnVectorValues.DocIndexIterator asDocIndexIterator(IndexedDISI disi) {
+    // can we replace with fromDISI?
+    return new KnnVectorValues.DocIndexIterator() {
+      @Override
+      public int docID() {
+        return disi.docID();
+      }
+
+      @Override
+      public int index() {
+        return disi.index();
+      }
+
+      @Override
+      public int nextDoc() throws IOException {
+        return disi.nextDoc();
+      }
+
+      @Override
+      public int advance(int target) throws IOException {
+        return disi.advance(target);
+      }
+
+      @Override
+      public long cost() {
+        return disi.cost();
+      }
+    };
+  }
 
   @Override
   public int docID() {
