+++ lucene/core/src/java/org/apache/lucene/codecs/lucene99/OffHeapQuantizedByteVectorValues.java
@@ -287,7 +287,7 @@ public abstract class OffHeapQuantizedByteVectorValues extends QuantizedByteVect
 
     @Override
     public DocIndexIterator createIterator() {
-      return DocIndexIterator.fromIndexedDISI(disi);
+      return IndexedDISI.asDocIndexIterator(disi);
     }
 
     @Override
